move protected methods to authorization.py
