+++ backend/ops_api/ops/resources/notifications.py
@@ -14,7 +14,8 @@ from sqlalchemy.orm import InstrumentedAttribute
 from typing_extensions import override
 
 from models import Notification, OpsEventType, User
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.utils.events import OpsEventHandler
 from ops_api.ops.utils.query_helpers import QueryHelper
