+++ backend/ops_api/ops/resources/portfolios.py
@@ -4,7 +4,8 @@ from typing_extensions import Any, List, override
 
 from models import Portfolio
 from models.base import BaseModel
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.utils.response import make_response_with_headers
