+++ backend/ops_api/ops/resources/procurement_steps.py
@@ -18,7 +18,9 @@ from models.workflows import (
     ProcurementStep,
     Solicitation,
 )
-from ops_api.ops.auth.auth import ExtraCheckError, Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
+from ops_api.ops.auth.exceptions import ExtraCheckError
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.schemas.procurement_steps import (
     AcquisitionPlanningRequest,
