+++ backend/ops_api/ops/resources/product_service_code.py
@@ -4,7 +4,8 @@ from flask import Response
 from typing_extensions import override
 
 from models.base import BaseModel
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
