+++ backend/ops_api/ops/resources/research_project_funding_summary.py
@@ -3,7 +3,8 @@ from flask import Response, request
 from typing_extensions import override
 
 from models.base import BaseModel
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import BaseListAPI, handle_api_error
 from ops_api.ops.utils.research_project_helper import (
     GetResearchProjectFundingSummaryQueryParams,
