+++ backend/ops_api/ops/resources/research_type.py
@@ -2,7 +2,8 @@ from flask import Response
 from typing_extensions import override
 
 from models.projects import ResearchType
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import EnumListAPI, handle_api_error
