+++ backend/ops_api/ops/resources/services_component.py
@@ -9,7 +9,9 @@ from typing_extensions import override
 
 from models import OpsEventType, ServicesComponent
 from models.base import BaseModel
-from ops_api.ops.auth.auth import ExtraCheckError, Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
+from ops_api.ops.auth.exceptions import ExtraCheckError
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.schemas.services_component import (
     PATCHRequestBody,
