+++ backend/ops_api/ops/resources/workflow_approve.py
@@ -17,7 +17,8 @@ from models.workflows import (
     WorkflowStepInstance,
     WorkflowStepStatus,
 )
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, handle_api_error
 from ops_api.ops.utils.procurement_workflow_helper import create_procurement_workflow
 from ops_api.ops.utils.response import make_response_with_headers
