+++ backend/ops_api/ops/resources/workflow_instance.py
@@ -17,7 +17,8 @@ from models.workflows import (
     WorkflowStepStatus,
     WorkflowTriggerType,
 )
-from ops_api.ops.auth.auth import Permission, PermissionType, is_authorized
+from ops_api.ops.auth.auth_enum import Permission, PermissionType
+from ops_api.ops.auth.authorization import is_authorized
 from ops_api.ops.base_views import BaseItemAPI, BaseListAPI, handle_api_error
 from ops_api.ops.resources.workflow_step_instance import WorkflowStepInstanceResponse
